confctl: Add filter_objects and update_objects
In some cases, we need to select conftool objects to act upon based not
only on their tags, but also on their value.

A typical example of this: we want to depool only services that were
previously pooled, do some action, then repool only those.

Create two methods that allow to fetch such objects, and to update them.

Change-Id: I5c40d51906700a76bf3e8b7df400aa4bfd6ee363