Add the LBRemoteCluster class.
For clusters behind a load balancer, we typically want to be able to
depool a server from a specific service, then run any number of commands
on it, and finally repool it.

We also want to ensure we can only have at max N hosts depooled at any
time. Given cumin doesn't have pre- and post- execution hooks, a special
run() method is created that breaks the remote run in smaller groups and
executes them all in parallel. Note this works a bit differently than
how the cumin moving window works, as here we'll have to wait for the
execution on all servers in a group before moving on to the next.

Change-Id: I303dbd3ba87cefb22ae2dad3a92d12785af6aa23