elasticsearch: Store which dcs to query in class
The cookbooks repo which relies upon this code will now expose a
--write-queue-datacenter flag to allow the user to specify one or mores dc to
check rather than all of them; this flag should be used to override when
there's a problem with the metrics in a given datacenter, not used
routinely.

This patch makes the ncessary changes on the spicerack side of things to
facilitate that new flag in the cookbooks repo.

Bug: T261239
Change-Id: I0147a638b1d001e4ca25b97002968fe87dfc8107