decorators: Add an optional custom failure message to @retry. Use it in dnsdisc.
We've seen that the default message can cause confusion -- for example,
"Failed to call 'dnsdisc.check_ttl'" sounds like it means there was a
failure attempting to check the TTL, but it actually just means the TTL
hasn't updated yet, so we'll wait and retry. And that's a fairly normal
use case: in general, errors caught by @retry aren't permanent failures
but transient states -- that's why we use @retry in the first place.

There's probably no single generic message that covers all uses here.
Thus, this patch keeps the existing message as a default, but allows
functions to specify their own message as an argument to @retry.

Change-Id: Iec47b192589196a24e49500a8f9ec38e135c23ea