decorators: Log chained exception messages in @retry
The @retry decorator logs the messages from exceptions raised during
execution, but when there are chained exceptions ("raise from", etc.)
only the top-level error is logged. For example, in
MediaWiki._check_siteinfo, we only log "Failed to get siteinfo" and
throw away the message from the underlying RequestException.

Instead, this traverses the exception chain (using the same logic as the
built-in default handler for uncaught exceptions) and includes each
exception's message in the log entry.

(In passing: Some elasticsearch tests failed because
TransportError.__str__ fails (!) if the exception isn't initialized with
enough arguments. This adds args to all TransportError constructed in
the test; it didn't matter before since we never called __str__.)

Change-Id: I54f50ea77b1ae24c03a0087f7fc2379b43b5c9a7