tox: Remove '--skip B322' from Bandit config.
B322 was the check that blacklists input(). We were skipping it because
input() is fine to use in Python 3.

Now that Bandit 1.7.0 is Python 3 only, they removed the check, and
'--skip B322' errors out (instead of quietly doing nothing) which caused
tox to fail every patch. This removes the flag so that Jenkins passes
again.

Change-Id: Ib120c5f357d3a0098e5378187a51d2239e1601c4