remote: allow prepending every command with sudo
This allows to ensure that every command run on the remote will be
prefixed by sudo (in a very primitive way).
This becomes really useful when you have to pass the RemoteHosts
object around to libraries to run other commands, that still require
sudo to run (but don't do so themselves).

Bug: T267412
Change-Id: I7ec3b4c25721c4f3d52ec2217451ea342d9897f2
Signed-off-by: David Caro <david@dcaro.es>