wait_reboot_since() is now using a constant backoff.
Previously, a linear backoff with a base delay of 10 seconds was
used. Since we do expect the reboot of a server to take some time,
by the time the server has rebooted, the retry interval has already
grown to multiple minutes. Since we don't expect more than one
process running this check, a constant backoff should be appropriate
and should increase the reactivity of this check significantly.

Change-Id: Ic430ef1d34903830bd98f82aa436c504a556e796