mediawiki: Update cronjob code now that most are systemd timers
Most of the MediaWiki "crons" have actually been migrated to systemd
timers, which gives us much nicer options to stop/start them.

For disabling we:
* Disable all systemd timers so no new jobs start
* Stop all jobs to take care of existing processes
* Check `systemd is-enabled` to verify everything was in fact disabled.

To enable, we don't do anything as once it's re-enabled, puppet will
re-enable all the disabled timers.

One major change is that the disabling step now happens in both
datacenters as we want to make sure that the currently running jobs are
stopped and the jobs won't start until we explicitly want them to (by
re-enabling puppet).

Since I already made breaking changes to the type signatures of some of
the functions, I went ahead and renamed all instance of "cronjob" that
don't refer to actual cronjobs to be "periodic job" as a generic term
that covers both cronjobs and systemd timers.

Bug: T266717
Change-Id: I311667d0584e4dd9a0101981c61d1786147271e6