mediawiki: Remove unnecessary and broken disable of systemd timers
Originally I had thought that the first `systemctl stop ...` command would
only stop the running services and we needed to explicitly disable the
timers, but that's not the case. `systemctl stop ...` stops both the
services and timers.

At which point the second disable command errors out because the first
`systemctl list-units ..` command finds nothing. Just remove it since
it's not necesssary.

Bug: T285524
Change-Id: I31bc72aa59235472fd9671fd1aeabf1711e2665f