ganeti: Use --force in shutdown
The decommission cookbook runs the ganeti.shutdown() function of Spicerack
to shutdown a VM. If an VM however is in ADMIN_down state (which was the
case for a test VM for the MX migration that we once created with the
makevm cookbook, but never installed), gnt-instance fails to shut down
properly. In addition the --timeout parameter isn't honored; we pass 2 mins
in the Spicerack function, but the decommission cookbook was still stuck
after the interval has passed.

Use --force with the gnt-instance shutdown command. It also forces the
shutdown for offline instances and should have no downside for machines
which are running. It confirmed that it's working fine by unblocking
the VM above.

Change-Id: I1e3388f58f4eb483bed0af251cc1ca6f47dca1a3