elastic: Restart masters one at a time after all others
Elastic strongly suggests, starting somewhere in the 7.x releases (but
probably previously as well, just not as obviously) that cluster
upgrades restart the master capable instances last. Only instances
with a version >= the master can join the cluster. If the master were
upgraded and an old node is partitioned from the cluster during an
upgrade, or if the upgrade has to be paused, the old instance will not
be allowed to re-join the cluster.

Because we restart one row at a time, and we are configured to always
be able to continue running if a full row is lost, this has an indirect
guarantee that enough masters remain in the cluster at all times.

Bug: T306389
Change-Id: Iba72c138c338c85d10addd4e76c5d5ae86d9de1a