elasticsearch: Simplify routine to start masters last
The routine to decide which nodes can be restarted was too strict. It
achieves the goal of restarting masters once all the nodes in the
cluster are restarted, but our production setup is configured such
that node A is master in cluster X and child in cluster Y, while node
B is master in cluster Y and child in cluster X, meaning each node
would have to be restarted before the other.

Instead simplify things away. The issue being solved is that
non-masters need to be restarted first. If the only nodes remaining to
be restarted are master capable, return them all as restartable.
Otherwise filter all the master capable nodes out of the restart list.

Change-Id: Ia70ade2a00796dfaee3ec5e3beac7dc171c3b1fa