alertmanager: Add support for multiple alertmanager instances
Adds support for using multiple alertmanager instances, based on a a
configuration file. One of those instances can be marked as 'default'
which is used when the call to the Spicerack.alertmanager() or
Spicerack.alertmanager_hosts() API is used without specifying a specific
instance or some other API (like Service.downtime()) that does not
support multiple instances is used.

The current implementation is not fully usable for the metricsinfra use
case quite yet, as that will need support for using a HTTP proxy and
likely some form of authentication. I will send a follow-up patch to
implement those assuming the direction this is going gets a +1.

Bug: T360932
Change-Id: Ib652e7c7ee4b7984da0492f286371c9d343da8d3