redfish: add the add_account function
Supermicro ships their servers with the BMC admin account set to
ADMIN, meanwhile we standardized the usage of root inside Wikimedia
(basically what Dell does by default).
This change adds a new add_account function that uses Redfish to
create a new account, that we'll use in the provision cookbook to
create the root account on new Supermicro servers.

I chose to use a simple API vs something more elaborated (for example,
to automatically set a new password if the user that is being added
is already existing etc..) since the code became quickly less
maintainable in my opinion.

The add_account function has been added only for Supermicro since
after various tests, the DELL BMCs seem not to accept any POST
parameter for the same URI. We don't need this use case for the DELL
vendor, so it is easier to just add the new functionality
for Supermicro.

Bug: T365372
Change-Id: I0eb4e0c4a0f09bd205229172f065f74606a84293