dhcp: add dhcp_filename and dhcp_options for DHCPConfMac and DHCPConfOpt82
We need to be able to set DHCP settings like "filename"
and various "options" in order to set TFTP-only DHCP configurations.

Add the possibility to specify the DHCP "filename" and "options XXXX"
settings for DHCP with Opt 82 and with Mac address.

This work may also be useful in the future if we'll test EFI.

Add also some extra whitelist settings for the new propertis to avoid
vulture's failures reported in prospector runs.

Bug: T363576
Change-Id: Id3f6046ceff65a438acf4e793e5c9ab3761f4cde