dhcp: allow empty distro for DHCPConfMac and DHCPConfOpt82
Allow "distro" to be empty, so that the correspondent pathprefix
config is not rendered. This is useful when we want to add
DHCP configs for IP configuration only, like the Supermicro
BMC/mgmt interface.

Note: I haven't modified DHCPConfMgmt since it seems very specific
and hopefully in the future the mac-address will not be neeed.
Allowing an empty distro value seems to add flexibility as well
to our configurations.

Bug: T365372
Change-Id: I8b5bb35e6a5354f5b16e2bad6125d7787578c592