redfish: catch no-json-responses in change_user_password
The Supermicro's Redfish implementation works the same as Dell's
in change_user_password, except for the fact that no JSON response
is returned.
Add a try/except block around it to avoid unnecessary failures
when dealing with Supermicro's BMC. It seems better than
copy/paste the code everywhere and create duplications.

Bug: T365372
Change-Id: Iaec81499a4bc8852105b56310241635258634f14