dhcp: Add option to omit sending filename to a vendor
Prior to this commit we always send the filename as part of the DHCP
offer. After this commit the filename can be omitted for the specified
vendor.

This functionality is needed to omit the filename when responding to the
debian installer.  Per Debian's "B.2.6. Using a DHCP server to specify
preconfiguration files" install doc a filename that looks like an HTTP
URI is loaded as a preseed configuration. This causes an issue when we
supply an UEFI HTTP Boot URL, as the debian installer fails when trying
to load the boot file as a preseed config. Since we don't supply
preseeds via DHCP, return an empty preseed to the debian installer

Change-Id: I980b30199e8a6230cd1c784c901a7ae921671ac4