redfish: add response logging for request()
It happened to me several times to not be able to inspect
the HTTP response's json/text for a 400 error code becase
the "response" payload is gone when raising a generic
RedfishError from request().

Attempt to log the JSON payload of a response when
the error code is between 400 and 600.

Fix also the pydoc to reflect the current behavior.

Change-Id: I22fda07716dd76d1e04499910bfc2f097f944fda