Spicerack: find true physical int if server primary IP is on a bridge
Moritz recently hit this error when trying to reimage ganeti2027:

https://phabricator.wikimedia.org/P71851

The reason is that the primary IP of the ganeti device is placed on
the 'private' bridge device on the system, and imported to Netbox the
same.  The current spicerack logic thus fails as it finds the
interface the primary IP is on and then tries to find what its
connected to.

This patch adds additional logic that if the primary device IP address
is connected to a bridge it does another lookup to find the physical
interface that belongs to the bridge.

Bug: T383207
Change-Id: I3baca91d678be3baa16b794eaecf01d3468c7703