icinga: skip downtimed services in wait_for_optimal if needed
We need to be able to instruct wait_for_optimal() about what
service checks can be skipped, because we don't want to wait
for their completion (for example, because it is expected that they
raise a warning or an error).

Add a specific option in wait_for_optimal() to skip downtimed services,
like we do for the acked ones, and add the related tests.

Bug: T392848
Change-Id: Ie4435f66956798cb006970b02d9c806d1a0a0798