redfish: add support for iDRAC 10 to force_http_boot_once
On new Dells with iDRAC 10 the current URI used is read only,
meanwhile a newer one can be used appending "/Settings".
The only drawback is that BootSourceOverrideMode is read only
even in /Settings, so it needs to be removed.

Tested the change with sretest2006 in UEFI mode:
1) Used the /Settings URI with the reduced boot options.
2) Performed a chassis reset and verified that Pxe was selected.
3) Performed another chassis reset and verified regular boot.

The only info about BootSourceOverrideMode that I found
(beside trying to figure out why it is read only in Redfish)
is https://github.com/ansible-collections/community.general/issues/5091
(TL;DR IIUC is that the option is not meant to be set with UEFI,
but it is not documented).

Bug: T393044
Bug: T392851
Change-Id: Ia60bee4ac65d1d50d2825219cd229eb29fc75d58