redfish: simplify change_user_password for Supermicro
In recent Supermicro models the "Username" attribute is read-only
and cannot be modified. I tested the API without "Username"
on sretest2001 and it worked nicely, so it makes sense to avoid
any extra ifs and just use the version that works in all use cases.

The "Username" attribute is indeed not needed since we find the right
User API path using find_account() earlier on.

Bug: T396365
Change-Id: I063b876ea75cb1e9aa1d3d2b2b92859f896ba825