redfish: support weak Etag values in change_user_password
On IDRAC 10 hosts the /Accounts Redfish endpoint (used by the module
in find_account(), in turn used by change_user_password()) doesn't
return anymore a regular Etag, but a weak one, that cannot be used
in the HTTP If-Match header since it causes validation failures.

Add a use case to change_user_password to skip the If-Match HTTP
header when a weak Etag is returned by Redfish in the users' data.

More details in T392851#11151819

Bug: T392851
Change-Id: I407f96518f0121ea6acea962cc514fefd2ca5aec