redfish: improve log_entries for idrac 10
For some reason that I don't explain yet, on IDRAC 10 the firmware
version starts at a much lower value, like 1.2.x rather than 3.3.x
or 4.x.x, but the APIs are the most recent ones anyway.

As example the log_entries property currently returns:

/redfish/v1/Managers/Logs/Lclog

But this leads to a HTTP 404 on IDRAC 10, meanwhile the more
canonical path (the one pointed out by super()) works
perfectly.

This impacts other calls based on this like last_reboot,
that in turn limits the update-firmware cookbook.

Bug: T392851
Change-Id: I9c615bc027e15dba6c1826c0d9e8a662efff1548